test(sdk-next): move cross-package contract identity out of client
packages/client removed its cyclic core/server dev dependencies, so its
contract-identity test only resolved through stale node_modules links and
failed on fresh installs. Keep the schema-only assertions in client and move
the Core/Server identity and generation-parity assertions to sdk-next, which
composes Client, Core, and Server.
